# Customer-Order-Portal/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart On X Customer Ordering Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portal/orders:
    get:
      operationId: listPortalOrders
      tags: [portal]
      summary: List orders (role-scoped) with filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: locationId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /portal/orders/{id}:
    get:
      operationId: getPortalOrder
      tags: [portal]
      summary: Get order detail (role-scoped)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found or outside scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Submit a new order (wizard submit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreatedResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── PORTAL ACCOUNT ──────────────────────────────────────────────
...
  /portal/account/emails:
    post:
      operationId: addAccountEmail
      tags: [portal]
      summary: Add an alias email to current user's contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Email added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── PORTAL LOCATIONS ─────────────────────────────────────────────
...
  /portal/location-requests:
    get:
      operationId: listPortalLocationRequests
      tags: [portal]
      summary: List location requests by current user's org
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationRequest"
    post:
      operationId: createLocationRequest
      tags: [portal]
      summary: Submit a new location request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationRequest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── PORTAL MANAGE TEAM ──────────────────────────────────────────
...
  /portal/manage/contacts/{id}/emails:
    post:
      operationId: addContactEmail
      tags: [portal]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Email added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── PORTAL MANAGE LOCATIONS ─────────────────────────────────────
  /portal/manage/locations:
    get:
      operationId: listManageLocations
      tags: [portal]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HsLocation"
    post:
      operationId: createManageLocation
      tags: [portal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsLocation"
  /portal/manage/locations/{id}:
    patch:
      operationId: updateManageLocation
      tags: [portal]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsLocation"
...
  /admin/organizations/{id}/locations:
    get:
      operationId: listAdminOrgLocations
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HsLocation"
    post:
      operationId: createAdminOrgLocation
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsLocation"
  /admin/organizations/{id}/locations/{locId}:
    patch:
      operationId: updateAdminOrgLocation
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: locId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsLocation"
...
  /admin/organizations/{id}/orders:
    get:
      operationId: listAdminOrgOrders
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

  # ─── ADMIN ONBOARDING ─────────────────────────────────────────────
...
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      parameters:
        - name: orgId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: terms
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
...
  /admin/orders/{id}:
    patch:
      operationId: updateAdminOrder
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"

  # ─── ADMIN LOCATION REQUESTS ─────────────────────────────────────
  /admin/location-requests:
    get:
      operationId: listAdminLocationRequests
      tags: [admin]
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationRequest"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── AUTH SCHEMAS ─────────────────────────────────────────────
...
    EmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
...
    HsBillingProfile:
      type: object
      required: [id, companyId, name, address1, city, state, zip, apEmail, paymentTerms, isDefault, validated]
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        apEmail:
          type: string
        apPhone:
          type: ["string", "null"]
        paymentTerms:
          type: string
        netDays:
          type: ["integer", "null"]
        isDefault:
          type: boolean
        validated:
          type: boolean
        placeId:
          type: ["string", "null"]
        usedByLocationCount:
          type: number
    HsLocation:
      type: object
      required: [id, companyId, billingProfileId, nickname, address1, city, state, zip, isDefault, validated]
      properties:
        id:
          type: string
        companyId:
          type: string
        billingProfileId:
          type: string
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: ["string", "null"]
        isDefault:
          type: boolean
        validated:
          type: boolean
        placeId:
          type: ["string", "null"]
        validationSource:
          type: string
        billingProfile:
          $ref: "#/components/schemas/HsBillingProfile"
...
    OrderItem:
      type: object
      required: [id, productId, skuSnapshot, nameSnapshot, quantity, unitPriceCents, lineTotalCents]
      properties:
        id:
          type: number
        productId:
          type: number
        skuSnapshot:
          type: string
        nameSnapshot:
          type: string
        quantity:
          type: number
        unitPriceCents:
          type: number
        lineTotalCents:
          type: number
    OrderSummary:
      type: object
      required: [id, orderNumber, hubspotCompanyId, hubspotLocationId, poNumber, status, paymentStatus, subtotalCents, shippingCents, totalCents, orderedAt]
      properties:
        id:
          type: number
        orderNumber:
          type: string
        hubspotCompanyId:
          type: string
        hubspotLocationId:
          type: string
        hubspotContactId:
          type: string
        portalUserId:
          type: number
        placedByRole:
          type: string
        poNumber:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        subtotalCents:
          type: number
        shippingCents:
          type: number
        totalCents:
          type: number
        orderedAt:
          type: string
          format: date-time
        shopifyOrderNumber:
          type: ["string", "null"]
        hubspotPoId:
          type: ["string", "null"]
        hubspotPoStatus:
          type: string
        locationNickname:
          type: ["string", "null"]
        orderedByUsername:
          type: ["string", "null"]
        orgName:
          type: ["string", "null"]
        itemCount:
          type: number
    OrderDetail:
      type: object
      required: [id, orderNumber, poNumber, status, paymentStatus, subtotalCents, shippingCents, totalCents, orderedAt, shipSnapshot, billingSnapshot, paymentTermsSnapshot, items]
      properties:
        id:
          type: number
        orderNumber:
          type: string
        hubspotCompanyId:
          type: string
        hubspotLocationId:
          type: string
        hubspotContactId:
          type: string
        portalUserId:
          type: number
        placedByRole:
          type: string
        poNumber:
          type: string
        orderNotes:
          type: ["string", "null"]
        status:
          type: string
        paymentStatus:
          type: string
        subtotalCents:
          type: number
        shippingCents:
          type: number
        totalCents:
          type: number
        orderedAt:
          type: string
          format: date-time
        shipSnapshot:
          type: object
          additionalProperties: true
        billingSnapshot:
          type: object
          additionalProperties: true
        paymentTermsSnapshot:
          type: string
        netDaysSnapshot:
          type: ["integer", "null"]
        shopifyOrderId:
          type: ["string", "null"]
        shopifyOrderNumber:
          type: ["string", "null"]
        hubspotPoId:
          type: ["string", "null"]
        hubspotPoStatus:
          type: string
        orderedByUsername:
          type: ["string", "null"]
        orgName:
          type: ["string", "null"]
        locationNickname:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderListResponse:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderSummary"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: number
        quantity:
          type: number
          minimum: 1
    OrderInput:
      type: object
      required: [hubspotLocationId, billingProfileId, poNumber, items]
      properties:
        hubspotLocationId:
          type: string
        billingProfileId:
          type: string
        poNumber:
          type: string
          maxLength: 40
        orderNotes:
          type: string
          maxLength: 500
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
          minItems: 1
        mockCardName:
          type: ["string", "null"]
        mockCardNumber:
          type: ["string", "null"]
        mockCardExpiry:
          type: ["string", "null"]
        mockCardCvc:
          type: ["string", "null"]
    OrderCreatedResponse:
      type: object
      required: [orderId, orderNumber]
      properties:
        orderId:
          type: number
        orderNumber:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string

    # ─── DASHBOARD SCHEMAS ───────────────────────────────────────
...
    LocationRequest:
      type: object
      required: [id, hubspotCompanyId, requestedByUserId, nickname, address1, city, state, zip, status, createdAt]
      properties:
        id:
          type: number
        hubspotCompanyId:
          type: string
        requestedByUserId:
          type: number
        requestedByUsername:
          type: ["string", "null"]
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: ["string", "null"]
        validated:
          type: boolean
        placeId:
          type: ["string", "null"]
        status:
          type: string
        resolvedBy:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        orgName:
          type: ["string", "null"]
    LocationRequestInput:
      type: object
      required: [nickname, address1, city, state, zip]
      properties:
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
          minLength: 2
          maxLength: 2
        zip:
          type: string
        phone:
          type: ["string", "null"]
        placeId:
          type: ["string", "null"]
        validated:
          type: boolean
...
    LocationInput:
      type: object
      required: [nickname, address1, city, state, zip, billingProfileId]
      properties:
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: ["string", "null"]
        billingProfileId:
          type: string
        isDefault:
          type: boolean
        placeId:
          type: ["string", "null"]
        validated:
          type: boolean
    LocationUpdate:
      type: object
      properties:
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: ["string", "null"]
        billingProfileId:
          type: string
        isDefault:
          type: boolean
        placeId:
          type: ["string", "null"]
        validated:
          type: boolean

    # ─── BILLING SCHEMAS ─────────────────────────────────────────
...